requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
requests>=2.26.0
certifi>=2023.0.0
